Break the single netcdf logic file -> several files
