+++ src/utilities/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_subdirectory(aerodynamics)
 add_subdirectory(controllers)
-add_subdirectory(outputs)
+add_subdirectory(netcdf)
