+++ src/utilities/netcdf/CMakeLists.txt
@@ -2,5 +2,5 @@ target_sources(openturbine_library PRIVATE)
 
 install(FILES
   netcdf_utilities.hpp
-  DESTINATION include/OpenTurbine/utilities/outputs/
+  DESTINATION include/OpenTurbine/utilities/netcdf/
 )
