+++ tests/unit_tests/utilities/CMakeLists.txt
@@ -1,7 +1,7 @@
-# Specify the source files for the unit test executable
+add_subdirectory(netcdf)
+
 target_sources(
     openturbine_unit_tests
     PRIVATE
-    test_netcdf_utilities.cpp
     test_turbine_parsing.cpp
 )
